# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApplyCheckIn": {
        "additionalProperties": false,
        "properties": {
          "actorReference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actorreference"
          },
          "currentText": {
            "title": "Currenttext",
            "type": "string"
          },
          "editorId": {
            "title": "Editorid",
            "type": "string"
          },
          "revision": {
            "default": "",
            "title": "Revision",
            "type": "string"
          },
          "targetReference": {
            "title": "Targetreference",
            "type": "string"
          }
        },
        "required": [
          "editorId",
          "currentText",
          "targetReference"
        ],
        "title": "ApplyCheckIn",
        "type": "object"
      },
...
      "ContextAttachmentIn": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "provenance": {
            "default": "caller",
            "title": "Provenance",
            "type": "string"
          }
        },
        "required": [
          "category",
          "content"
        ],
        "title": "ContextAttachmentIn",
        "type": "object"
      },
      "ContextPreviewIn": {
        "additionalProperties": false,
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/ContextAttachmentIn"
            },
            "title": "Attachments",
            "type": "array"
          },
          "databaseVersion": {
            "default": "",
            "title": "Databaseversion",
            "type": "string"
          },
          "schema": {
            "default": "",
            "title": "Schema",
            "type": "string"
          },
          "targetReference": {
            "title": "Targetreference",
            "type": "string"
          }
        },
        "required": [
          "targetReference"
        ],
        "title": "ContextPreviewIn",
        "type": "object"
      },
      "CopilotRequestIn": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actorIsDurable": {
            "default": true,
            "title": "Actorisdurable",
            "type": "boolean"
          },
          "actorReference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actorreference"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/ContextAttachmentIn"
            },
            "title": "Attachments",
            "type": "array"
          },
          "conversationId": {
            "default": "",
            "title": "Conversationid",
            "type": "string"
          },
          "databaseVersion": {
            "default": "",
            "title": "Databaseversion",
            "type": "string"
          },
          "editor": {
            "$ref": "#/components/schemas/EditorIn"
          },
          "protocolVersion": {
            "default": "1.0",
            "title": "Protocolversion",
            "type": "string"
          },
          "schema": {
            "default": "",
            "title": "Schema",
            "type": "string"
          },
          "targetReference": {
            "title": "Targetreference",
            "type": "string"
          },
          "userMessage": {
            "default": "",
            "title": "Usermessage",
            "type": "string"
          }
        },
        "required": [
          "action",
          "targetReference"
        ],
        "title": "CopilotRequestIn",
        "type": "object"
      },
...
      "EditorIn": {
        "additionalProperties": false,
        "properties": {
          "editorId": {
            "default": "",
            "title": "Editorid",
            "type": "string"
          },
          "revision": {
            "default": "",
            "title": "Revision",
            "type": "string"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          }
        },
        "title": "EditorIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "OracleDBHarness is the shared execution, policy and copilot layer for Oracle work.\n\nEvery operation carries an operation ID, an actor, a target, typed parameters,\ncapability requirements, a risk class and limits, and produces an execution record\nwith a structured result and verification evidence.\n\nNotes that apply to the whole API:\n\n* Free-form SQL runs only on targets where an administrator enabled worksheets. A\n  keyword filter is not a read-only boundary, so the Oracle account is the boundary.\n* Production targets are observation only in this release.\n* Applying a copilot proposal changes an editor buffer. It never runs, compiles or\n  commits anything.\n",
    "title": "OracleDBHarness API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/copilot/context/preview": {
      "post": {
        "description": "Show the user exactly what would be sent, before anything is sent.",
        "operationId": "context_preview_api_v1_copilot_context_preview_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextPreviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Context Preview Api V1 Copilot Context Preview Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Context Preview",
        "tags": [
          "copilot"
        ]
      }
    },
...
    "/api/v1/copilot/proposals/{proposal_id}/apply-check": {
      "post": {
        "description": "Decide whether a proposal may still be applied to this buffer.\n\nApplying updates the editor only. It never runs, compiles, commits or pushes\nanything.",
        "operationId": "apply_check_api_v1_copilot_proposals__proposal_id__apply_check_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyCheckIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Apply Check Api V1 Copilot Proposals  Proposal Id  Apply Check Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Check",
        "tags": [
          "copilot"
        ]
      }
    },
    "/api/v1/copilot/requests": {
      "post": {
        "operationId": "copilot_request_api_v1_copilot_requests_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopilotRequestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Copilot Request",
        "tags": [
          "copilot"
        ]
      }
    },
...
    "/api/v1/integrations/capabilities": {
      "get": {
        "description": "What this harness supports. Carries no secrets and takes no source text.",
        "operationId": "integration_capabilities_api_v1_integrations_capabilities_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Integration Capabilities Api V1 Integrations Capabilities Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Integration Capabilities",
        "tags": [
          "copilot"
        ]
      }
    },
...
  }
}